refactor: FetchPost 내부 로직 분리
